feat(FE): useScrollDirection 훅 제작 및 커스텀 훅 내보내기 방식 변경
- useMediaQuery 훅을 기본 내보내기로 변경
- `shared/hooks/index.ts`에서 모든 훅의 내보내기를 관리하고, 관련된 모든 import 경로 수정
- useScrollDirection 훅을 사용하여 PageTemplate 및 Header 컴포넌트의 스크롤 방향 로직 단순화
- PageTemplate 및 Header 컴포넌트의 스타일 및 클래스 이름 최적화
