+++ front/src/shared/hooks/useScrollDirection.tsx
@@ -1,0 +1,23 @@
+import { useState, useEffect } from "react";
+
+export default function useScrollDirection() {
+  const [isScrollingUp, setIsScrollingUp] = useState(true);
+  const [lastScrollY, setLastScrollY] = useState(0);
+
+  useEffect(() => {
+    const handleScroll = () => {
+      const currentScrollY = window.scrollY;
+      if (currentScrollY < lastScrollY) {
+        setIsScrollingUp(true);
+      } else if (currentScrollY > lastScrollY) {
+        setIsScrollingUp(false);
+      }
+      setLastScrollY(currentScrollY);
+    };
+
+    window.addEventListener("scroll", handleScroll, { passive: true });
+    return () => window.removeEventListener("scroll", handleScroll);
+  }, [lastScrollY]);
+
+  return isScrollingUp;
+}
